move data factory param function
