+++ stacks/data/generate/templates/processing/processing_template_DQ/spark_jobs/data_quality.py.jinja
@@ -1,5 +1,6 @@
 import logging
 
+from stacks.data.azure.data_factory import get_data_factory_param
 from stacks.data.constants import CONFIG_CONTAINER_NAME
 from stacks.data.pyspark.data_quality.main import data_quality_main
 from stacks.data.logger import setup_logger
